+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Added a view as table for array and object values in `struct` view
 - Added `destroy()` method to Popup view
 - Added `external` data option for `badge` views
 - Fixed a crush on open actions popup in `struct` view when a value has too much nested structure